# Project1.0/src/main/java/com/ecommerce/product/controller/ProductApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ecommerce.product.controller;

import com.ecommerce.product.model.Product;
import com.ecommerce.product.model.ProductResponse;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-06-01T10:11:09.516412300+05:30[Asia/Calcutta]")

@Validated
@Api(value = "product", description = "the product API")
public interface ProductApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /product/{sellerID}/{ProductName} : Delete product
     * This is a delete product.
     *
     * @param productName The name that needs to be deleted (required)
     * @param sellerID  (required)
     * @return operation successful (status code 200)
     */
    @ApiOperation(value = "Delete product", nickname = "deleteProduct", notes = "This is a delete product.", tags={ "SellerProduct", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "operation successful") })
    @RequestMapping(value = "/product/{sellerID}/{ProductName}",
        method = RequestMethod.DELETE)
    default ResponseEntity<Void> deleteProduct(@ApiParam(value = "The name that needs to be deleted",required=true) @PathVariable("ProductName") String productName,@ApiParam(value = "",required=true) @PathVariable("sellerID") String sellerID) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /product/{category} : get the list of specific category of products
     * It gets all the products from the specific category
     *
     * @param category  (required)
     * @param productname  (optional)
     * @param pageNumber  (optional, default to 0)
     * @param pageSize  (optional, default to 15)
     * @return okay accepted (status code 200)
     */
    @ApiOperation(value = "get the list of specific category of products", nickname = "getAllCategoryProducts", notes = "It gets all the products from the specific category", response = ProductResponse.class, tags={ "UserProduct","SellerProduct", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "okay accepted", response = ProductResponse.class) })
    @RequestMapping(value = "/product/{category}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<ProductResponse> getAllCategoryProducts(@ApiParam(value = "",required=true) @PathVariable("category") String category,@ApiParam(value = "") @Valid @RequestParam(value = "productname", required = false) String productname,@ApiParam(value = "", defaultValue = "0") @Valid @RequestParam(value = "pageNumber", required = false, defaultValue="0") Integer pageNumber,@ApiParam(value = "", defaultValue = "15") @Valid @RequestParam(value = "pageSize", required = false, defaultValue="15") Integer pageSize) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"result\" : [ { \"quantity\" : 10, \"sellerNumber\" : \"ae04a3db-8169-4f91-95f9-478bdb107840\", \"price\" : 12.25, \"name\" : \"shoes\", \"discount\" : 20, \"details\" : \"description of the image that we want to show\", \"category\" : \"hardware\" }, { \"quantity\" : 10, \"sellerNumber\" : \"ae04a3db-8169-4f91-95f9-478bdb107840\", \"price\" : 12.25, \"name\" : \"shoes\", \"discount\" : 20, \"details\" : \"description of the image that we want to show\", \"category\" : \"hardware\" } ], \"statusMessage\" : \"Operation completed successful\", \"statusCode\" : 200 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /product/seller/{sellerID}/{category} : get the list of specific seller selling products
     * It gets all the products from the specific seller
     *
     * @param category  (required)
     * @param sellerID  (required)
     * @param pageNumber  (optional, default to 0)
     * @param pageSize  (optional, default to 15)
     * @return okay accepted (status code 200)
     */
    @ApiOperation(value = "get the list of specific seller selling products", nickname = "getAllSellerProducts", notes = "It gets all the products from the specific seller", response = ProductResponse.class, tags={ "SellerProduct", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "okay accepted", response = ProductResponse.class) })
    @RequestMapping(value = "/product/seller/{sellerID}/{category}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<ProductResponse> getAllSellerProducts(@ApiParam(value = "",required=true) @PathVariable("category") String category,@ApiParam(value = "",required=true) @PathVariable("sellerID") String sellerID,@ApiParam(value = "", defaultValue = "0") @Valid @RequestParam(value = "pageNumber", required = false, defaultValue="0") Integer pageNumber,@ApiParam(value = "", defaultValue = "15") @Valid @RequestParam(value = "pageSize", required = false, defaultValue="15") Integer pageSize) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"result\" : [ { \"quantity\" : 10, \"sellerNumber\" : \"ae04a3db-8169-4f91-95f9-478bdb107840\", \"price\" : 12.25, \"name\" : \"shoes\", \"discount\" : 20, \"details\" : \"description of the image that we want to show\", \"category\" : \"hardware\" }, { \"quantity\" : 10, \"sellerNumber\" : \"ae04a3db-8169-4f91-95f9-478bdb107840\", \"price\" : 12.25, \"name\" : \"shoes\", \"discount\" : 20, \"details\" : \"description of the image that we want to show\", \"category\" : \"hardware\" } ], \"statusMessage\" : \"Operation completed successful\", \"statusCode\" : 200 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /product/{sellerID}/{ProductName} : get a single product from the database by its name
     * returns a single product by matching name
     *
     * @param productName  (required)
     * @param sellerID  (required)
     * @return okay accepted (status code 200)
     */
    @ApiOperation(value = "get a single product from the database by its name", nickname = "getSingleProductBySeller", notes = "returns a single product by matching name", response = Product.class, tags={ "SellerProduct", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "okay accepted", response = Product.class) })
    @RequestMapping(value = "/product/{sellerID}/{ProductName}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Product> getSingleProductBySeller(@ApiParam(value = "",required=true) @PathVariable("ProductName") String productName,@ApiParam(value = "",required=true) @PathVariable("sellerID") String sellerID) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"quantity\" : 10, \"sellerNumber\" : \"ae04a3db-8169-4f91-95f9-478bdb107840\", \"price\" : 12.25, \"name\" : \"shoes\", \"discount\" : 20, \"details\" : \"description of the image that we want to show\", \"category\" : \"hardware\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /product/productView/{ProductName} : get a single product from the database by its name
     * returns a single product by matching name
     *
     * @param productName  (required)
     * @return okay accepted (status code 200)
     */
    @ApiOperation(value = "get a single product from the database by its name", nickname = "getSingleProducts", notes = "returns a single product by matching name", response = Product.class, tags={ "UserProduct","OrderProduct", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "okay accepted", response = Product.class) })
    @RequestMapping(value = "/product/productView/{ProductName}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Product> getSingleProducts(@ApiParam(value = "",required=true) @PathVariable("ProductName") String productName) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"quantity\" : 10, \"sellerNumber\" : \"ae04a3db-8169-4f91-95f9-478bdb107840\", \"price\" : 12.25, \"name\" : \"shoes\", \"discount\" : 20, \"details\" : \"description of the image that we want to show\", \"category\" : \"hardware\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /product/{sellerID}/{ProductName} : changethe details of single product from the database by its name
     * change  a single product by matching name
     *
     * @param productName  (required)
     * @param sellerID  (required)
     * @param product  (required)
     * @return okay accepted (status code 200)
     */
    @ApiOperation(value = "changethe details of single product from the database by its name", nickname = "updateSingleProducts", notes = "change  a single product by matching name", response = Product.class, tags={ "SellerProduct", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "okay accepted", response = Product.class) })
    @RequestMapping(value = "/product/{sellerID}/{ProductName}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    default ResponseEntity<Product> updateSingleProducts(@ApiParam(value = "",required=true) @PathVariable("ProductName") String productName,@ApiParam(value = "",required=true) @PathVariable("sellerID") String sellerID,@ApiParam(value = "" ,required=true )  @Valid @RequestBody Product product) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"quantity\" : 10, \"sellerNumber\" : \"ae04a3db-8169-4f91-95f9-478bdb107840\", \"price\" : 12.25, \"name\" : \"shoes\", \"discount\" : 20, \"details\" : \"description of the image that we want to show\", \"category\" : \"hardware\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}